final syntax edits, now testing
